requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-astradb>=0.1.0
astrapy>=1.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
openai>=1.10.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
Jinja2>=3.1.0
